--- app/src/main/res/drawable/ic_referrals.xml
@@ -1,3 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:width="24dp" android:height="24dp" android:viewportWidth="24" android:viewportHeight="24">
-    <path android:pathData="M18.88 5.9c1.11 0 2.04 0.772 2.197 1.795l0.017 0.148 0.006 0.152v8.01c0 1.05-0.821 1.926-1.902 2.073l-0.157 0.017L18.88 18.1H6.12c-1.11 0-2.04-0.772-2.197-1.795l-0.017-0.148L3.9 16.005v-8.01c0-1.05 0.821-1.926 1.902-2.073L5.96 5.906 6.12 5.9h12.76zm-3.501 6.81l-1.772 1.837c-0.53 0.548-1.388 0.615-2 0.183l-0.119-0.093-0.094-0.09-1.775-1.837-4.129 4.275 0.01 0.006c0.112 0.063 0.234 0.108 0.361 0.134l0.129 0.019 0.13 0.006h12.76c0.213 0 0.418-0.051 0.598-0.147l0.029-0.018-4.128-4.275zm4.71-4.875l-4.042 4.183 4.037 4.179 0.012-0.07 0.007-0.122v-8.01l-0.006-0.112-0.008-0.048zm-15.179 0L4.903 7.876l-0.006 0.12v8.008c0 0.041 0.002 0.082 0.007 0.122l0.011 0.07 4.036-4.178L4.91 7.835zM18.88 6.85H6.12c-0.22 0-0.431 0.055-0.616 0.156L5.468 7.028l6.633 6.867c0.171 0.177 0.441 0.216 0.65 0.11l0.086-0.054 0.062-0.056 6.632-6.867-0.035-0.022C19.357 6.93 19.204 6.88 19.043 6.86L18.88 6.85z" android:fillColor="?colorPrimary" android:fillType="evenOdd"/>
-</vector>
