+++ testdata/src/main/java/com/hedvig/app/testdata/feature/referrals/Data.kt
@@ -2,12 +2,14 @@ package com.hedvig.app.testdata.feature.referrals
 
 import com.hedvig.android.owldroid.fragment.MonetaryAmountFragment
 import com.hedvig.android.owldroid.fragment.ReferralFragment
+import com.hedvig.android.owldroid.type.Feature
 import com.hedvig.app.testdata.common.builders.CostBuilder
 import com.hedvig.app.testdata.feature.referrals.builders.EditCodeDataBuilder
 import com.hedvig.app.testdata.feature.referrals.builders.LoggedInDataBuilder
 import com.hedvig.app.testdata.feature.referrals.builders.ReferralsDataBuilder
 
 val LOGGED_IN_DATA_WITH_REFERRALS_FEATURE_ENABLED = LoggedInDataBuilder().build()
+val LOGGED_IN_DATA_WITH_REFERRALS_FEATURE_DISABLED = LoggedInDataBuilder(listOf(Feature.KEYGEAR)).build()
 val LOGGED_IN_DATA_WITH_KEY_GEAR_DISABLED = LoggedInDataBuilder(features = listOf()).build()
 
 val REFERRALS_DATA_WITH_NO_DISCOUNTS = ReferralsDataBuilder().build()
